Guard COLR paint recursion per-offset and unwind painter state on throw
The PaintDecycler was entered only at PaintGlyph / PaintColrGlyph, so
every other recursive edge - transforms, composites, layers - recursed
through PaintParser.TryParse unguarded. A PaintTransform whose sub-paint
offset points at itself, or any long acyclic transform chain, recursed
into an uncatchable StackOverflow from rendering one glyph of a hostile
font. Re-key the decycler on the paint offset (uint) and enter it at the
top of TryParse, so every edge is bounded by both cycle detection and
the depth limit; the offset key subsumes the old glyph-boundary check
(each glyph resolves to one base-paint offset), which is removed.

ColorGlyphV1Painter pushed transform/clip/layer states onto the caller's
DrawingContext, popped only by the traverser's non-finally calls. A
throw mid-paint (a failed outline build, a draw error) skipped the pop
and leaked the pushed states into the caller's context, corrupting
later drawing. The painter now owns a single LIFO state stack and is
IDisposable; Draw disposes it via `using`, so an escaping throw unwinds
whatever was pushed. Folding the two stacks into one also removes a
latent out-of-order disposal hazard.

Also: compute ColrLayers layer-offset arithmetic in long (a hostile
firstLayerIndex / paint offset could wrap an int/uint and slip the
bounds check), and drop two dead PaintDecycler rents in
ColorGlyphV1Drawing (the resolved paint tree is walked without one).

+++ src/Avalonia.Base/Media/Fonts/Tables/Colr/PaintDecycler.cs
@@ -3,9 +3,15 @@
 namespace Avalonia.Media.Fonts.Tables.Colr
 {
     /// <summary>
-    /// Type alias for the paint decycler that tracks visited glyphs.
+    /// Decycler for the COLR v1 paint graph, keyed on the absolute paint offset within the COLR
+    /// table. Every recursive parse edge goes through <see cref="PaintParser.TryParse"/>, which
+    /// enters this guard on the paint's offset — so both paint→paint cycles (an offset that
+    /// reappears on the current path) and unbounded acyclic nesting (depth past
+    /// <see cref="MaxTraversalDepth"/>) are caught, mirroring HarfBuzz's per-paint nesting cap.
+    /// Offsets uniquely identify a paint (and each glyph resolves to one base-paint offset), so an
+    /// offset key also subsumes the older glyph-boundary cycle check.
     /// </summary>
-    internal class PaintDecycler : Decycler<ushort>
+    internal class PaintDecycler : Decycler<uint>
     {
         /// <summary>
         /// Maximum depth for paint graph traversal.
